libqmi-glib: use ((guint64) 1) << 31 for QMI_*_BAND_CAPABILITY_BC_15
The current definition of QMI_*_BAND_CAPABILITY_BC_15 gets converted to a
negative value (0xffffffff80000000) which causes false positives for BC15
to be returned from dms_add_qmi_bands() and nas_add_qmi_bands() in
ModemManager/src/mm-modem-helpers-qmi.c when a matching QMI band (e.g.
WCDMA900) is present in qmi_bands like in this example
https://lists.freedesktop.org/archives/libqmi-devel/2016-March/001572.html .

Replace 1 << 31  with ((guint64) 1) << 31 for QMI_*_BAND_CAPABILITY_BC_15
to avoid incorrect mmcli "Bands | supported: 'cdma-bc15-aws, ...'" output.

(cherry picked from commit 0f7849ce05f3e26283652c20e5392161c1cfeeaf)

+++ src/libqmi-glib/qmi-flags64-dms.h
@@ -111,7 +111,7 @@ typedef enum {
     QMI_DMS_BAND_CAPABILITY_BC_12            = 1 << 28,
     QMI_DMS_BAND_CAPABILITY_BC_14            = 1 << 29,
     /* Bit 30 reserved */
-    QMI_DMS_BAND_CAPABILITY_BC_15            = 1 << 31,
+    QMI_DMS_BAND_CAPABILITY_BC_15            = ((guint64) 1) << 31,
     /* Bits 32-47 reserved */
     QMI_DMS_BAND_CAPABILITY_WCDMA_2600       = ((guint64) 1) << 48,
     QMI_DMS_BAND_CAPABILITY_WCDMA_900        = ((guint64) 1) << 49,
